+++ dogfightZ/src/graphic_Z/HUDs/CharLabel.java
@@ -2,7 +2,7 @@ package graphic_Z.HUDs;
 
 public class CharLabel extends CharHUD
 {
-	protected short location[];
+	protected int location[];
 	protected String text;
 	//protected char fraps_buffer[][];	//帧缓冲区(引用，实体在CharVisualManager中)
 	
@@ -11,8 +11,8 @@ public class CharLabel extends CharHUD
 	public CharLabel
 	(
 		char frapsBuffer[][], 
-		short HUDLayer, 
-		short scrResolution[],
+		int HUDLayer, 
+		int scrResolution[],
 		boolean transparentAtSpace
 	)
 	{
@@ -20,7 +20,7 @@ public class CharLabel extends CharHUD
 		resolution = scrResolution;
 		layer		 = HUDLayer;
 		fraps_buffer = frapsBuffer;
-		location = new short[2];
+		location = new int[2];
 		text = new String("");
 		location[0] = 0;
 		location[1] = 0;
@@ -31,11 +31,11 @@ public class CharLabel extends CharHUD
 	public CharLabel
 	(
 		char frapsBuffer[][], 
-		short HUDLayer, 
-		short scrResolution[], 
+		int HUDLayer, 
+		int scrResolution[], 
 		String Text, 
-		short location_X, 
-		short location_Y,
+		int location_X, 
+		int location_Y,
 		boolean transparentAtSpace
 	)
 	{
@@ -49,11 +49,11 @@ public class CharLabel extends CharHUD
 	public CharLabel
 	(
 		char frapsBuffer[][], 
-		short HUDLayer, 
-		short scrResolution[], 
+		int HUDLayer, 
+		int scrResolution[], 
 		String Text, 
-		short location_X, 
-		short location_Y
+		int location_X, 
+		int location_Y
 	)
 	{
 		this(frapsBuffer, HUDLayer, scrResolution, Text, location_X, location_Y, true);
@@ -64,18 +64,14 @@ public class CharLabel extends CharHUD
 		if(visible)
 		{
 			char tmp;
-			for
-			(
-				short x=location[0], i=0, j=0; 
+			for (
+				int x=location[0], i=0, j=0; 
 				x<resolution[0] && i<text.length(); 
 				++i, ++x
-			)	if ((tmp=(char)text.charAt(i)) != ' ' || transparentAtSpace == false)
-			{
-				if(tmp == '\n')
-				{
-					++j; x=(short) (location[0]-1);
-				}
-				else fraps_buffer[location[1] + j][x] = tmp;
+			)	if ((tmp=(char)text.charAt(i)) != ' ' || transparentAtSpace == false) {
+				if(tmp == '\n') {
+					++j; x = location[0] - 1;
+				} else fraps_buffer[location[1] + j][x] = tmp;
 			}
 		}
 	}
@@ -83,9 +79,9 @@ public class CharLabel extends CharHUD
 	public CharLabel
 	(
 		char frapsBuffer[][], 
-		short HUDLayer, 
-		short scrResolution[], 
-		short X, short Y
+		int HUDLayer, 
+		int scrResolution[], 
+		int X, int Y
 	)
 	{
 		this(frapsBuffer, HUDLayer, scrResolution, true);
@@ -98,7 +94,7 @@ public class CharLabel extends CharHUD
 		text = new String(Text);
 	}
 	
-	public void setLocation(short X, short Y)
+	public void setLocation(int X, int Y)
 	{
 		location[0] = X;
 		location[1] = Y;
